:sparkles: carousel refactor
