+++ src/components/Carousel/components/Dots.tsx
