+++ src/components/Carousel/components/SliderContent.tsx
