+++ src/components/Carousel/components/sliderImage.ts
@@ -1,0 +1,27 @@
+export default [
+  {
+    title: 'Lorem Ipsum',
+    description: 'Lorem Ipsum',
+    urls: 'Lorem Ipsum',
+  },
+  {
+    title: 'Lorem Ipsum',
+    description: 'Lorem Ipsum',
+    urls: 'Lorem Ipsum',
+  },
+  {
+    title: 'Lorem Ipsum',
+    description: 'Lorem Ipsum',
+    urls: 'Lorem Ipsum',
+  },
+  {
+    title: 'Lorem Ipsum',
+    description: 'Lorem Ipsum',
+    urls: 'Lorem Ipsum',
+  },
+  {
+    title: 'Lorem Ipsum',
+    description: 'Lorem Ipsum',
+    urls: 'Lorem Ipsum',
+  },
+]
