+++ src/components/Carousel/slider.css
@@ -91,36 +91,26 @@
   height: 100%;
   position: absolute;
   display: flex;
-  top: 92%;
+  top: 96%;
   justify-content: center;
   z-index: 200;
 }
 
 .dot {
   cursor: pointer;
-  height: 1.5rem;
-  width: 1.5rem;
-  margin: 0px 3px;
-  background-color: transparent;
+  height: 1rem;
+  width: 1rem;
+  margin: 0px 4px;
+
   background-color: rgba(0, 0, 0, 0.3);
   border-radius: 50%;
   display: inline-block;
   
 }
 
-.slider-container:hover .dot:hover {
-  background-color: rgba(255, 255, 255, 0.5);
-}
 
-.active-dot {
-  background-color: rgba(255, 255, 255, 0.5);
-}
-
-.slider-container:hover .dot{
-  background-color: rgba(0, 0, 0, 0.3);
-}
-.slider-container:hover .active-dot{
-  background-color: rgba(255, 255, 255, 0.5);
+.slider-container .active-dot{
+  background-color: rgba(0, 0, 0, 0.5);
 }
 
 .play-pause {
