--- src/components/Carousel/sliderImage.ts
@@ -1,32 +1,0 @@
-export default [
-  {
-    title: 'Comece a jogar agora!',
-    description:
-      'A versão DEMO contém as regras básicas para você começar a jogar 3DeT Victory. Com este material você pode criar seus personagens...',
-    urls: 'https://s3-sa-east-1.amazonaws.com/cdn.br.catarse/uploads/redactor_rails/picture/data/501849/1ad74bfb-94e5-42ae-9e57-aac2fa20b79c.png',
-  },
-  {
-    title: 'Apoie e receba as regras completas',
-    description:
-      'A versão playtest com todas as regras do Manual 3DeT Victory já está disponível para apoiadores dos Níveis 1, 2, 3 e 4!',
-    urls: 'https://s3-sa-east-1.amazonaws.com/cdn.br.catarse/uploads/redactor_rails/picture/data/501943/0ef151d1-e613-4834-9a29-c0792b602225.png',
-  },
-  {
-    title: 'ADD-ONS',
-    description:
-      'Prepare-se! Toda semana tem novidades aqui no Catarse de 3DeT Victory. Vamos oferecer livros com preços e condições especiais...',
-    urls: 'https://s3-sa-east-1.amazonaws.com/cdn.br.catarse/uploads/redactor_rails/picture/data/500458/e6b33e34-e1fe-436c-bbe1-97ec0c33686a.png',
-  },
-  {
-    title: 'Kit Dados das Arcas',
-    description:
-      'Aventure-se pelas Arcas com os dados oficiais de 3DeT! São 3 dados coloridos de seis lados...',
-    urls: 'https://s3-sa-east-1.amazonaws.com/cdn.br.catarse/uploads/redactor_rails/picture/data/502815/e5309d28-3516-4cf8-9fa4-d7ef404e38d4.png',
-  },
-  {
-    title: 'Defensores de Tóquio',
-    description:
-      'Lançada em 1995 nas bancas de jornal, Defensores de Tóquio marcou gerações de jogadores...',
-    urls: 'https://s3-sa-east-1.amazonaws.com/cdn.br.catarse/uploads/redactor_rails/picture/data/498441/a0cb3cf8-fd2b-46a5-8d1e-635e492785db.png',
-  },
-]
